[Attention] Mamba attention module refactor - LINEAR (#43556)
